+++ module/VuFind/config/module.config.php
@@ -445,7 +445,7 @@ $config = [
             'VuFind\ChannelProvider\ChannelLoader' => 'VuFind\ChannelProvider\ChannelLoaderFactory',
             'VuFind\ChannelProvider\PluginManager' => 'VuFind\ServiceManager\AbstractPluginManagerFactory',
             'VuFind\Condition\Manager' => 'VuFind\ServiceManager\Factory\AutowiringFactory',
-            'VuFind\Condition\Handlers\PluginManager' => 'VuFind\ServiceManager\AbstractPluginManagerFactory',
+            'VuFind\Condition\Handler\PluginManager' => 'VuFind\ServiceManager\AbstractPluginManagerFactory',
             'VuFind\Config\AccountCapabilities' => 'VuFind\Config\AccountCapabilitiesFactory',
             'VuFind\Config\ConfigManager' => 'VuFind\Config\ConfigManagerFactory',
             'VuFind\Config\Upgrade' => 'VuFind\Config\UpgradeFactory',
