+++ module/VuFind/src/VuFind/Condition/Handler/Date.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * String condition handler (used for tests).
+ * Date condition handler.
  *
  * PHP version 8
  *
@@ -22,28 +22,26 @@
  * <https://www.gnu.org/licenses/>.
  *
  * @category VuFind
- * @package  Condition_Handlers
+ * @package  Condition_Handler
  * @author   Nathan Collins <colli372@msu.edu>
  * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
 
-namespace VuFind\Condition\Handlers;
-
-use VuFind\Exception\ConditionException;
+namespace VuFind\Condition\Handler;
 
 /**
- * String condition handler (used for tests).
+ * Date condition handler.
  *
  * @category VuFind
- * @package  Condition_Handlers
+ * @package  Condition_Handler
  * @author   Nathan Collins <colli372@msu.edu>
  * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
-class StringHandler extends AbstractBase
+class Date extends AbstractBase
 {
     /**
      * Get base value to check.
@@ -51,16 +49,9 @@ class StringHandler extends AbstractBase
      * @param array $condition Optional used for handler specific parameters
      *
      * @return string
-     * @throws ConditionException
      */
     protected function getBaseValue(array $condition): string
     {
-        $string = $condition['string'] ?? null;
-        if ($string === null) {
-            throw new ConditionException(
-                'String condition handler requires key "string" specifying the value to check.'
-            );
-        }
-        return $string;
+        return (new \DateTime())->format('Y-m-d');
     }
 }
