--- module/VuFind/src/VuFind/Condition/Handlers/PluginManager.php
@@ -1,75 +1,0 @@
-<?php
-
-/**
- * Condition handler plugin manager
- *
- * PHP version 8
- *
- * Copyright (C) Hebis Verbundzentrale 2026.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, see
- * <https://www.gnu.org/licenses/>.
- *
- * @category VuFind
- * @package  Condition_Handlers
- * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://vufind.org/wiki/development Wiki
- */
-
-namespace VuFind\Condition\Handlers;
-
-use Laminas\ServiceManager\Factory\InvokableFactory;
-
-/**
- * Condition handler plugin manager
- *
- * @category VuFind
- * @package  Condition_Handlers
- * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://vufind.org/wiki/development Wiki
- */
-class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
-{
-    /**
-     * Default plugin aliases.
-     *
-     * @var array
-     */
-    protected $aliases = [
-        'date' => Date::class,
-        'string' => StringHandler::class,
-    ];
-
-    /**
-     * Default plugin factories.
-     *
-     * @var array
-     */
-    protected $factories = [
-        Date::class => InvokableFactory::class,
-        StringHandler::class => InvokableFactory::class,
-    ];
-
-    /**
-     * Return the name of the base class or interface that plug-ins must conform
-     * to.
-     *
-     * @return string
-     */
-    protected function getExpectedInterface()
-    {
-        return ConditionHandlerInterface::class;
-    }
-}
